requirements.txt
watchdog
websockets
fastapi
uvicorn
pydantic
openai
httpx
twilio
python-dotenv